Migrate date_range_tests into its own script
